feat: add marker pins

+++ app/Data/PositionData.php
@@ -10,7 +10,7 @@ class PositionData extends Data
 {
     public function __construct(
         //
-        public int $x,
-        public int $y
+        public float $x,
+        public float $y
     ) {}
 }
